specification in the wild
