+++ catalyst/rl/offpolicy/sampler.py
@@ -1,16 +1,20 @@
+from typing import List
+
 import os
 import time
 import copy
 import random
 from datetime import datetime
 import torch
+import torch.nn as nn
 from tensorboardX import SummaryWriter
 
 from catalyst.utils.misc import set_global_seed
 from catalyst.dl.utils import UtilsFactory
-from catalyst.utils.serialization import serialize, deserialize
 from catalyst.rl.offpolicy.utils import EnvWrapper
 from catalyst.rl.offpolicy.exploration import ExplorationHandler
+from catalyst.rl.environments.core import EnvironmentSpec
+from catalyst.rl.db.core import DBSpec
 
 # speed up optimization
 os.environ["OMP_NUM_THREADS"] = "1"
@@ -22,21 +26,18 @@ _SEED_RANGE = _BIG_NUM
 class Sampler:
     def __init__(
         self,
-        network,
-        env,
-        id,
-        logdir=None,
-        redis_server=None,
-        redis_prefix=None,
-        buffer_size=int(1e4),
-        weights_sync_period=1,
-        mode="infer",
-        resume=None,
-        seeds=None,
-        action_clip=None,
-        episode_limit=None,
-        force_store=False,
-        discrete_actions=False
+        network: nn.Module,  # @TODO: actor policy specification?
+        env: EnvironmentSpec,
+        db_server: DBSpec = None,
+        logdir: str = None,
+        id: int = 0,
+        mode: str = "infer",
+        buffer_size: int = int(1e4),
+        weights_sync_period: int = 1,
+        seeds: List = None,
+        episode_limit: int = None,
+        force_store: bool = False,
+        resume: str = None,
     ):
 
         self._seed = 42 + id
@@ -44,26 +45,13 @@ class Sampler:
         self._sampler_id = id
 
         # logging
-        if logdir is not None:
-            current_date = datetime.now().strftime("%y-%m-%d-%H-%M-%S-%M-%f")
-            logpath = f"{logdir}/sampler-{mode}-{id}-{current_date}"
-            os.makedirs(logpath, exist_ok=True)
-            self.logger = SummaryWriter(logpath)
-        else:
-            self.logger = None
-
-        # main attributes
-        self.weights_sync_period = weights_sync_period
+        self._prepare_logger(logdir, mode)
 
         # synchronization configuration
-        self.redis_server = redis_server
-        self.redis_prefix = redis_prefix or ""
+        self.db_server = db_server
+        self.weights_sync_period = weights_sync_period
         self.episode_limit = episode_limit or _BIG_NUM
         self._force_store = force_store
-
-        # discrete_actions = isinstance(self.env.action_space, Discrete)
-        # self._sampler_weight_mode = \
-        #     "critic" if discrete_actions else "actor"
 
         # other
         self._infer = mode == "infer"
@@ -82,24 +70,32 @@ class Sampler:
             deterministic=self._infer
         )
 
+        self._sampler_weight_mode = \
+            "critic" if env.discrete_actions else "actor"
+
         # resume
         if resume is not None:
             self.load_checkpoint(resume=resume)
 
+    def _prepare_logger(self, logdir, mode):
+        if logdir is not None:
+            current_date = datetime.now().strftime("%y-%m-%d-%H-%M-%S-%M-%f")
+            logpath = f"{logdir}/sampler-{mode}-{id}-{current_date}"
+            os.makedirs(logpath, exist_ok=True)
+            self.logger = SummaryWriter(logpath)
+        else:
+            self.logger = None
+
     def _to_tensor(self, *args, **kwargs):
         return torch.Tensor(*args, **kwargs).to(self._device)
 
-    def load_checkpoint(self, *, resume=None, redis_server=None):
+    def load_checkpoint(self, *, resume: str = None, db_server: DBSpec = None):
         if resume is not None:
             checkpoint = UtilsFactory.load_checkpoint(resume)
             weights = checkpoint[f"{self._sampler_weight_mode}_state_dict"]
             self.network.load_state_dict(weights)
-        elif redis_server is not None:
-            weights = deserialize(
-                redis_server.get(
-                    f"{self.redis_prefix}_{self._sampler_weight_mode}_weights"
-                )
-            )
+        elif db_server is not None:
+            weights = db_server.load_weights(suffix=self._sampler_weight_mode)
             weights = {k: self._to_tensor(v) for k, v in weights.items()}
             self.network.load_state_dict(weights)
         else:
@@ -107,11 +103,10 @@ class Sampler:
         self.network.eval()
 
     def _store_trajectory(self):
-        if self.redis_server is None:
+        if self.db_server is None:
             return
         trajectory = self.env_wrapper.get_trajectory(tolist=True)
-        trajectory = serialize(trajectory)
-        self.redis_server.rpush("trajectories", trajectory)
+        self.db_server.push_trajectory(trajectory)
 
     def _prepare_seed(self):
         seed = self._seed + random.randrange(_SEED_RANGE)
@@ -159,7 +154,7 @@ class Sampler:
     def run(self):
         while True:
             if self.episode_index % self.weights_sync_period == 0:
-                self.load_checkpoint(redis_server=self.redis_server)
+                self.load_checkpoint(db_server=self.db_server)
             seed = self._prepare_seed()
             exploration_strategy = \
                 self.exploration_handler.get_exploration_strategy()
